+++ nacc_attribute_deriver/attributes/derived/modules/__init__.py
@@ -8,5 +8,5 @@ from .cross_module import *
 from .np import *
 from .mds import *
 from .meds import *
-from .milestone import *
+from .mlst import *
 from .uds import *