+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Documentation of release versions of `nacc-attribute-deriver`
 
 ## 2.2.0
 
+* Adds MLST missingness and changes scope to `mlst` not `milestone`
 * Update to include non-prescription V1 drugs (e.g. over-the-counter and vitamins - was previously only keeping track of prescription drugs)
 * Update to handle new B1a scope - like MEDS, comes from a separate file and has to be mapped to UDS
 * Updates to grab `BIRTHMO` and `BIRTHYR` from a pre-computed `subject.info.working.cross-sectional` location instead of files, to account for the fact that DOB can change across forms. The pre-computed value should be the latest